line_reply_service/requirements.txt
flask==2.3.3
requests>=2.28.1,<2.32
python-dotenv==1.0.1
flask_sqlalchemy==3.1.1
psycopg2-binary==2.9.9
flask_bcrypt==1.0.1
line-bot-sdk==2.4.1
Pillow==10.3.0
cachetools==5.3.3
